Creating AJAX

+++ foto/views.py
@@ -4,7 +4,7 @@ from django.shortcuts import render, get_object_or_404, redirect
 from django.db.models import Q, Count
 from .forms import FotoForm, CommentsForm, SubscribeForm, CommentsSecondLevelForm
 from .models import Foto, Category, Comments, User, CommentsSecondLevel
-from django.http import HttpResponseRedirect, HttpResponse
+from django.http import HttpResponseRedirect, HttpResponse, JsonResponse
 from django.urls import reverse
 from django.core.paginator import Paginator, EmptyPage
 
@@ -193,6 +193,10 @@ def edit_foto(request, foto_id):
         try:
             form = FotoForm(request.POST, instance=foto)
             form.save()
+            print(foto.images, request.POST)
+            if foto.images != request.POST.get('images', False):
+                foto.affected = False
+                foto.save()
             return redirect('user')
         except ValueError:
             return render(request, 'foto/add_foto.html', {'foto': foto, 'form': form, 'error': 'Bad data!'})
@@ -254,16 +258,21 @@ def edit_comment(request, comment_id):
 
 
 def delete_comment(request, comment_id):
-    comment = get_object_or_404(Comments, pk=comment_id, user=request.user)
-    comment.delete()
-    print('Deleted comment...')
-    return HttpResponse('Комментарий удален')
+    if request.method == 'GET':
+        comment = get_object_or_404(Comments, pk=comment_id, user=request.user)
+        print(request.GET)
+        comment.delete()
+        print('Deleted comment...')
+        #return redirect('foto', foto)
+        return redirect('home')
+
 
 def delete_comment_asinc(request):
-    comment = get_object_or_404(Comments, pk=request.GET.get('pk'), user=request.user)
-    comment.delete()
-    print('Deleted comment...')
-    return HttpResponse('Комментарий удален')
+    if request.method == 'DELETE':
+        print('Вошли в функцию')
+        comment = get_object_or_404(Comments, pk=request.GET.get('pk'), user=request.user)
+        comment.delete()
+        return HttpResponse('Комментарий удален')
 
 
 def add_comment_second_level(request,comment_id):
